- police-ratio added to players.xml - AI fleet count is set to log2 of (planet count + 2)

+++ open-ig/src/hu/openig/editors/AllocationPanel.java
@@ -234,7 +234,7 @@ public class AllocationPanel extends JPanel {
 			if (b.isReady()) {
 				totalWorkerDemand += b.getWorkers();
 				int e = b.getEnergy();
-				float eff = b.getEfficiency();
+				double eff = b.getEfficiency();
 				if (e < 0) {
 					totalEnergyDemand += e;
 				} else {
